+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- Windscribe provider. [#39](https://github.com/passepartoutvpn/passepartout-ios/issues/39)
+
 ## 1.2.0 (2019-04-01)
 
 ### Added